feat: remove webhooks, combine into single notifications app
