--- apps/marginfi-notifications/notifications-webhooks/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "@mrgnlabs/notifications-webhooks",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev --port 3005",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "firebase-admin": "^12.0.0",
-    "next": "14.1.0",
-    "react": "^18",
-    "react-dom": "^18",
-    "resend": "^3.1.0"
-  },
-  "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "eslint": "^8",
-    "eslint-config-next": "14.1.0",
-    "typescript": "^5"
-  }
-}
