--- apps/marginfi-notifications/notifications-webhooks/tsconfig.json
@@ -1,25 +1,0 @@
-{
-  "extends": "@mrgnlabs/tsconfig/nextjs.json",
-  "compilerOptions": {
-    "downlevelIteration": true,
-    "paths": {
-      "~/*": [
-        "./src/*"
-      ]
-    },
-    "plugins": [
-      {
-        "name": "next"
-      }
-    ]
-  },
-  "include": [
-    "next-env.d.ts",
-    "**/*.ts",
-    "**/*.tsx",
-    ".next/types/**/*.ts"
-  ],
-  "exclude": [
-    "node_modules"
-  ]
-}
