+++ apps/marginfi-notifications/package.json
