+++ apps/marginfi-notifications/src/config.ts
