+++ apps/marginfi-notifications/src/utils/index.ts
@@ -1,0 +1,11 @@
+export function sleep(ms: number) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
+
+export function getDebugLogger(context: string) {
+  return require("debug")(`mfi:liquidator:${context}`);
+}
+
+export function shortAddress(address: string) {
+  return `${address.slice(0, 6)}...${address.slice(-4)}`;
+}
